src/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
azure-cosmos